Make integer and float 32 bit

+++ fazic_mono/src/fazic/enums.rs
@@ -19,7 +19,7 @@ pub enum NodeElement {
 pub enum Value {
     Integer(i32),
     String(String),
-    Float(f64),
+    Float(f32),
     Bool(bool),
     Null,
 }
